captcha/requirements.txt
pillow==7.1.2
tensorflow==2.2.0
keras==2.4.0
opencv-python==4.2.0.34
imutils
scikit-learn
matplotlib